Add BroEater contract and related interfaces
This commit adds the BroEater contract, which implements the IEater interface and extends the Eater contract. The BroEater contract allows users to eat a specific token (underlying) by transferring the token from their account to the contract. It also includes functions to get and set the eat rate. Additionally, the commit adds the Eater contract, which is an abstract contract that provides the functionality to transfer tokens and emits an event when tokens are eaten. Finally, the commit adds the IEater interface, which defines the functions for eating tokens and getting/setting the eat rate.